requirements.txt
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
idna==3.4
importlib-resources==5.12.0
jsonpath-ng==1.5.3
jsonschema==4.17.3
lxml==4.9.2
pkgutil_resolve_name==1.3.10
ply==3.11
pyrsistent==0.19.3
python-dotenv==1.0.0
requests==2.28.2
robotframework==6.0.2
robotframework-jsonlibrary==0.5
six==1.16.0
urllib3==1.26.15
zipp==3.15.0
